# docs/swagger.yaml
basePath: /api/v1
definitions:
  golang-project-structure_internal_dto.MetaDTO:
    properties:
      currentPage:
        type: integer
      itemsPerPage:
        type: integer
      lastPage:
        type: integer
      totalItems:
        format: int64
        type: integer
    type: object
  golang-project-structure_internal_dto_user.CreateParams:
    properties:
      email:
        type: string
      name:
        type: string
    required:
    - email
    - name
    type: object
  golang-project-structure_internal_dto_user.ListOutput:
    properties:
      items:
        items:
          $ref: '#/definitions/golang-project-structure_internal_dto_user.UserDTO'
        type: array
      meta:
        $ref: '#/definitions/golang-project-structure_internal_dto.MetaDTO'
    type: object
  golang-project-structure_internal_dto_user.UpdateParams:
    properties:
      email:
        type: string
      name:
        type: string
    required:
    - email
    - name
    type: object
  golang-project-structure_internal_dto_user.UserDTO:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: HTTP API for the go-fx-clean-architecture service.
  title: go-fx-clean-architecture API
  version: "1.0"
paths:
  /users:
    get:
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/golang-project-structure_internal_dto_user.ListOutput'
      summary: List users
      tags:
      - users
    post:
      consumes:
      - application/json
      parameters:
      - description: User payload
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/golang-project-structure_internal_dto_user.CreateParams'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/golang-project-structure_internal_dto_user.UserDTO'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a user
      tags:
      - users
  /users/{id}:
    delete:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a user
      tags:
      - users
    get:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/golang-project-structure_internal_dto_user.UserDTO'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a user by ID
      tags:
      - users
    put:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: User payload
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/golang-project-structure_internal_dto_user.UpdateParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/golang-project-structure_internal_dto_user.UserDTO'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a user
      tags:
      - users
swagger: "2.0"